rrr: merge reactor cluster into single rrr.reactor module (clang 22)
Combines event.h+.cc, fiber_impl.h+.cc, quorum_event.h+.cc,
reactor.h+.cc, and fiber_context_runtime.cc into one C++23 named-
module interface unit at src/rrr/reactor/reactor.cpp (~2860 lines).
The cluster's classes (rrr::Event, rrr::Fiber, rrr::Reactor,
janus::QuorumEvent, rrr::fiber_task_t, …) form a mutually-recursive
web of forward declarations and out-of-line member definitions —
separate module interfaces would require circular import statements
which the C++23 standard forbids, so a single module unit is the
natural shape. The arch-specific context-switch trampolines stay
outside the module: fiber_context_x86_64.cc and
fiber_context_aarch64.cc are tiny `extern "C"` asm-only TUs and
don't need module attachment.

Unblocks the multi-attachment trap that blocked this merge on
clang 19 (see docs/dev/srpc_module_migration_plan.md): clang 22
correctly attaches implicit template instantiations to the
template's owning module (matching the C++23 standard), so
rusty::HashMap<std::string, …>'s shared RustyHash<std::string>
specialization no longer triggers "declaration X attached to named
module rrr.foo can't be attached to other modules" when consumers
import rrr.any_message + rrr.inmemory_channel + rrr.reactor
together.

Consumer changes (switching textual #include of reactor headers to
`import rrr.reactor;`):
  - src/rrr/rrr.hpp (umbrella)
  - src/rrr/rpc/{fiber_channel,tcp_channel,client,server}.hpp
  - src/rrr/reactor/{fiber,future}.h

Misc fixes uncovered by the merge:
  - src/rrr/misc/alarm.cpp switched from a GMF forward-decl of
    `PollThread` to `import rrr.reactor;` — clang 22 rejects the
    GMF forward-decl when another imported module exports the same
    class with full attachment.
  - src/rrr/rpc/tcp_channel.hpp added `#include <rusty/box.hpp>`;
    its `rusty::make_box` use was satisfied by reactor.h's
    transitive textual include, which the module BMI doesn't
    expose as a shim.
  - src/rrr/CMakeLists.txt wires `src/compat` BEFORE the third-
    party include path so the pre-existing
    `compat/rusty/function.hpp` shim adds the libc++22-required
    `#include <cstdlib>` ahead of upstream rusty/function.hpp. No
    third-party submodule patch needed.

Build with clang 22 (Homebrew 22.1.5):
  cmake -G Ninja -B build -S . -DCMAKE_BUILD_TYPE=Release \
    -DCMAKE_C_COMPILER=/home/users/shuai/.linuxbrew/bin/clang \
    -DCMAKE_CXX_COMPILER=/home/users/shuai/.linuxbrew/bin/clang++ \
    -DCMAKE_EXE_LINKER_FLAGS="-L/home/users/shuai/.linuxbrew/opt/llvm/lib \
      -Wl,-rpath,/home/users/shuai/.linuxbrew/opt/llvm/lib -stdlib=libc++" \
    -DCMAKE_SHARED_LINKER_FLAGS=…
  ninja -C build rrr rpcbench -j32

Clean build (rrr + rpcbench): 67s, librrr.a 12.7 MB (vs clang19's
83s / 9.7 MB pre-merge — both within noise).
